JWT vendor, JWT endpoint, coins endpoints, postmanJSON

+++ database/migrations/2019_01_20_225358_create_coin_historical_table.php
@@ -15,15 +15,15 @@ class CreateCoinHistoricalTable extends Migration
     {
     
         
-        Schema::create('coin_historical', function (Blueprint $table) {
-
+        Schema::create('coin_historical', function (Blueprint $table) 
+        {
             // we will add-> coin_id to know coin owner     
             //"price_usd": "2962.04162456",
             //"snapshot_at": "2018-04-03T10:54:02+00:00"
             
             $table->integer('coin_id')->unsigned()
                   ->foreign('coin_id')->references('id')->on('coin')->onDelete('cascade');
-            $table->float('price_usd')->nullable();
+            $table->double('price_usd')->nullable();
             $table->timestamp('snapshot_at');
             $table->timestamps();
         });
